+++ src/meltano/core/plugin/dbt/base.py
@@ -27,7 +27,8 @@ class DbtTransformPluginInstaller:
         self.project = project
         self.plugin = plugin
 
-    def install(self, reason):
+    async def install(self, reason):
+        """Install the transform into the project."""
         if reason in (PluginInstallReason.ADD, PluginInstallReason.UPGRADE):
             try:
                 transform_add_service = TransformAddService(self.project)